modal/requirements.txt
modal>=0.64.0
supabase>=2.0.0
httpx>=0.25.0
pycryptodome>=3.19.0
python-dotenv>=1.0.0
tenacity>=8.2.0
